add blockchain app
