+++ luce_vm/luce_django/luce/accounts/models.py
@@ -15,6 +15,9 @@ from django.conf import settings
 from brownie import project
 from brownie import network
 from brownie import accounts
+# from brownie.project.BrownieProject import PlonkVerifier
+# from blockchain.models import PlonkVerifierContract
+
 import urllib3
 import json
 from django.core.exceptions import ObjectDoesNotExist
@@ -27,36 +30,35 @@ import os
 logger = set_logger(__file__)
 
 # project should be loaded at somewhere else
-loaded_project = project.get_loaded_projects()
+# loaded_project = project.get_loaded_projects()
 
-if len(loaded_project) == 0:
-    print("No project")
-else:
-    for p in loaded_project:
-        p.close()
+# if len(loaded_project) == 0:
+#     print("No project")
+# else:
+#     for p in loaded_project:
+#         p.close()
 
-LUCE_PROJECT_PATH = os.path.join(settings.BASE_DIR, "../../brownie")
-# print(luce_project_path)
-print(LUCE_PROJECT_PATH)
-luce_project = project.load(LUCE_PROJECT_PATH)
-luce_project.load_config()
-# print("network.show_active()")
-print(network.is_connected())
-network.connect()
-print(network.is_connected())
+# LUCE_PROJECT_PATH = os.path.join(settings.BASE_DIR, "../../brownie")
+# # print(luce_project_path)
+# print(LUCE_PROJECT_PATH)
+# luce_project = project.load(LUCE_PROJECT_PATH)
+# luce_project.load_config()
+# print(luce_project)
+# # print("network.show_active()")
+# print(network.is_connected())
+# network.connect()
+# print(network.is_connected())
 
+# class Verifier(models.Model):
+#     address = models.CharField(max_length=255, null=True)
 
+#     def deploy(self):
+#         private_key = "56c6de6fd54438dbb31530f5fdeeaada0b1517880c91dfcd46a4e5fec59a9c79"
+#         new_account = accounts.add(private_key=private_key)
+#         contract = PlonkVerifier.deploy({'from': new_account})
 
-class Verifier(models.Model):
-    address = models.CharField(max_length=255, null=True)
-
-    def deploy(self):
-        private_key = "56c6de6fd54438dbb31530f5fdeeaada0b1517880c91dfcd46a4e5fec59a9c79"
-        new_account = accounts.add(private_key=private_key)
-        contract = luce_project.PlonkVerifier.deploy({'from': new_account})
-
-        self.address = contract.address
-        self.save()
+#         self.address = contract.address
+#         self.save()
 
 
 class UserManager(BaseUserManager):
@@ -266,400 +268,7 @@ class User(AbstractBaseUser):
         return self.active
 
 
-class Restrictions(models.Model):
-    no_restrictions = models.BooleanField()
-    open_to_general_research_and_clinical_care = models.BooleanField()
-    open_to_HMB_research = models.BooleanField()
-    open_to_population_and_ancestry_research = models.BooleanField()
-    open_to_disease_specific = models.BooleanField()
 
-
-class GeneralResearchPurpose(models.Model):
-    use_for_methods_development = models.BooleanField(default=False)
-    use_for_reference_or_control_material = models.BooleanField(default=False)
-    use_for_research_concerning_populations = models.BooleanField(
-        default=False)
-    use_for_research_ancestry = models.BooleanField(default=False)
-    use_for_biomedical_research = models.BooleanField(default=False)
-
-
-class HMBResearchPurpose(models.Model):
-    use_for_research_concerning_fundamental_biology = models.BooleanField(
-        default=False)
-    use_for_research_concerning_genetics = models.BooleanField(default=False)
-    use_for_research_concerning_drug_development = models.BooleanField(
-        default=False)
-    use_for_research_concerning_any_disease = models.BooleanField(
-        default=False)
-    use_for_research_concerning_age_categories = models.BooleanField(
-        default=False)
-    use_for_research_concerning_gender_categories = models.BooleanField(
-        default=False)
-
-
-class ClinicalPurpose(models.Model):
-    use_for_decision_support = models.BooleanField(default=False)
-    use_for_disease_support = models.BooleanField(default=False)
-
-
-class ResearchPurpose(models.Model):
-    general_research_purpose = models.ForeignKey(GeneralResearchPurpose,
-                                                 on_delete=models.CASCADE,
-                                                 null=True)
-    HMB_research_purpose = models.ForeignKey(HMBResearchPurpose,
-                                             on_delete=models.CASCADE,
-                                             null=True)
-    clinical_purpose = models.ForeignKey(ClinicalPurpose,
-                                         on_delete=models.CASCADE,
-                                         null=True)
-
-
-class ConsentContract(models.Model):
-    contract_address = models.CharField(max_length=255, null=True)
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    restrictions = models.ForeignKey(Restrictions, on_delete=models.CASCADE)
-    research_purpose = models.ForeignKey(ResearchPurpose,
-                                         on_delete=models.CASCADE,
-                                         null=True)
-
-    def update_data_consent(self):
-        transaction_dict = {
-            'from': accounts.add(private_key=self.user.ethereum_private_key)
-        }
-
-        logger.info(self.user.ethereum_public_key)
-        transaction_receipt = luce_project.ConsentCode.at(
-            self.contract_address).UploadDataPrimaryCategory(
-                self.user.ethereum_public_key,
-                self.restrictions.no_restrictions,
-                self.restrictions.open_to_general_research_and_clinical_care,
-                self.restrictions.open_to_HMB_research,
-                self.restrictions.open_to_population_and_ancestry_research,
-                self.restrictions.open_to_disease_specific, transaction_dict)
-
-        return transaction_receipt.status
-
-    def upload_data_consent(self, estimate):
-        return web3.upload_data_consent(self, estimate)
-
-    def retrieve_contract_owner(self):
-        return web3.retrieve_contract_owner(self)
-
-    def deploy(self):
-        private_key = self.user.ethereum_private_key
-        new_account = accounts.add(private_key=private_key)
-        contract = luce_project.ConsentCode.deploy({'from': new_account})
-
-        self.contract_address = contract.address
-        self.save()
-
-    def deploy_contract(self):
-        tx_receipt = web3.deploy_consent(self.user)
-        if type(tx_receipt) is list:
-            return tx_receipt
-        self.contract_address = tx_receipt["contractAddress"]
-        self.save()
-        return tx_receipt
-
-    def give_clinical_research_purpose(self, user, estimate):
-        rp = self.research_purpose.clinical_purpose
-        # print(rp)
-        logger.info(rp)
-        private_key = self.user.ethereum_private_key
-        new_account = accounts.add(private_key=private_key)
-
-        txn_dict = {'from': new_account}
-
-        receipt = luce_project.ConsentCode.at(
-            self.contract_address).giveClinicalPurpose(
-                user.ethereum_public_key,
-                rp.use_for_decision_support,
-                rp.use_for_disease_support,
-                txn_dict
-        )
-
-        logger.info(receipt)
-
-        return receipt.status
-
-        # tx = web3.give_clinical_research_purpose(self, user, estimate)
-        # return tx
-
-    def give_HMB_research_purpose(self, user, estimate):
-        rp = self.research_purpose.HMB_research_purpose
-        # print(rp)
-        logger.info(rp)
-        private_key = self.user.ethereum_private_key
-        new_account = accounts.add(private_key=private_key)
-
-        txn_dict = {'from': new_account}
-
-        receipt = luce_project.ConsentCode.at(
-            self.contract_address).giveHMBPurpose(
-                user.ethereum_public_key, 
-                rp.use_for_research_concerning_fundamental_biology,
-                rp.use_for_research_concerning_genetics,
-                rp.use_for_research_concerning_drug_development,
-                rp.use_for_research_concerning_any_disease,
-                rp.use_for_research_concerning_age_categories,
-                rp.use_for_research_concerning_gender_categories,
-                txn_dict
-        )
-
-        logger.info(receipt)
-        return receipt.status
-        # tx = web3.give_HMB_research_purpose(self, user, estimate)
-        # return tx
-
-    def give_general_research_purpose(self, user, estimate):
-        rp = self.research_purpose.general_research_purpose
-        private_key = self.user.ethereum_private_key
-        new_account = accounts.add(private_key=private_key)
-
-        txn_dict = {'from': new_account}
-
-        receipt = luce_project.ConsentCode.at(
-            self.contract_address).giveResearchPurpose(
-                user.ethereum_public_key, rp.use_for_methods_development,
-                rp.use_for_reference_or_control_material,
-                rp.use_for_research_concerning_populations,
-                rp.use_for_research_ancestry, rp.use_for_biomedical_research,
-                txn_dict)
-
-        # tx = web3.give_general_research_purpose(self, user, estimate)
-        logger.info(receipt)
-        return receipt.status
-
-
-class DataContract(models.Model):
-    contract_address = models.CharField(max_length=255, null=True, unique=True)
-
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-
-    commitment = models.CharField(max_length=255, null=False, default='0x0')
-
-    consent_contract = models.ForeignKey(ConsentContract,
-                                         on_delete=models.CASCADE,
-                                         null=True)
-
-    description = models.CharField(max_length=255, null=True)
-
-    licence = models.IntegerField(default=1)
-
-    link = models.CharField(max_length=255, null=True)
-
-    def get_a_new_account(self):
-        new_account = accounts.add()
-        accounts[0].transfer(new_account, 1e18)
-        return new_account
-
-    def require_verifier_deployed(self):
-        try:
-            v = Verifier.objects.get(pk=1)
-
-            if v.address is None:
-                deploy_result = v.deploy()
-                # print(deploy_result)
-
-                v.address = deploy_result.address
-                v.save()
-        except ObjectDoesNotExist:
-            v = Verifier.objects.create(pk=1)
-            r = v.deploy()
-
-            print(r)
-            # v.address = v.deploy().address
-            v.save()
-
-    def deploy(self):
-
-        # private_key = self.user.ethereum_private_key
-        # new_account = accounts.add(private_key=private_key)
-        new_account = self.get_a_new_account()
-        # accounts[1].transfer(new_account, 1e18)  # 1 ether
-        # print(new_account)
-
-        self.require_verifier_deployed()
-        verifier_address = Verifier.objects.get(pk=1).address
-        # print(verifier_address)
-
-        # commitment = self.get_commitment("hello")
-        commitment = {"public_signals":['16279978653553831575017442062517458639822344791369834134794885970235221444339']}
-        # commitment = "16279978653553831575017442062517458639822344791369834134794885970235221444339n"
-        print(commitment)
-
-        # print(accounts.at())
-        contract = luce_project.LuceMain.deploy(verifier_address,
-                                                commitment['public_signals'],
-                                                {'from': new_account})
-
-        self.contract_address = contract.address
-        self.save()
-        return contract.tx.txid
-
-    def get_commitment(self, secret):
-        http = urllib3.PoolManager()
-        snark_service_url = "http://zkp_service:8888/compute_commitment"
-        body_json = json.dumps({"secret": secret}).encode('utf-8')
-
-        print(body_json)
-
-        r = http.request('POST',
-                         snark_service_url,
-                         body=body_json,
-                         headers={'Content-Type': 'application/json'})
-
-        
-        result = json.loads(r.data.decode('utf-8'))
-
-        print(result)
-        return result
-
-    def deploy_contract(self):
-        # self.deploy()
-
-        tx_receipt = web3.deploy_contract_main(self.user)
-        if type(tx_receipt) is list:
-            return tx_receipt
-        self.contract_address = tx_receipt["contractAddress"]
-        self.save()
-        return tx_receipt
-
-    def set_registry_address(self, registry_address):
-        new_account = self.get_a_new_account()
-        transaction_dict = {'from': new_account}
-
-        transaction_receipt = luce_project.LuceMain.at(
-            self.contract_address).setRegistryAddress(registry_address,
-                                                      transaction_dict)
-
-        return transaction_receipt.status
-
-    # def set_registry_address(self, registry, estimate):
-    #     tx_receipt = web3.set_registry_address(self, registry.contract_address,
-    #                                            estimate)
-    #     return tx_receipt
-
-    def set_consent_address(self):
-        new_account = self.get_a_new_account()
-        transaction_dict = {'from': new_account}
-
-        transaction_receipt = luce_project.LuceMain.at(
-            self.contract_address).setConsentAddress(
-                self.consent_contract.contract_address, transaction_dict)
-
-        return transaction_receipt.status
-
-        # tx_receipt = web3.set_consent_address(
-        #     self, self.consent_contract.contract_address, estimate)
-        # return tx_receipt
-
-    def publish_dataset(self, link):
-        new_account = self.get_a_new_account()
-        transaction_dict = {
-            'from': new_account,
-        }
-
-        print(transaction_dict)
-
-        transaction_receipt = luce_project.LuceMain.at(
-            self.contract_address).publishData(self.description, link,
-                                               self.licence, transaction_dict)
-
-        return transaction_receipt.status
-        # tx = web3.publish_dataset(self, user, link)
-        # return tx
-
-    def retreive_info(self):
-        tx_receipt = web3.retreive_dataset_info(self)
-        return tx_receipt
-
-    def add_data_requester(self, access_time, purpose_code, user, estimate):
-        acc = accounts.add(private_key=user.ethereum_private_key)
-
-        txn_dict = {'from': acc, "gas_limit": 1e15, "allow_revert": True}
-
-        tx_receipt = luce_project.LuceMain.at(
-            self.contract_address).addDataRequester(purpose_code, access_time,
-                                                    txn_dict)
-
-        return tx_receipt
-
-    def getLink(self, user, estimate):
-        new_account = self.get_a_new_account()
-        transaction_dict = {'from': new_account}
-
-
-        
-        receipt = luce_project.LuceMain.at(self.contract_address).getLink(transaction_dict)
-        logger.info(receipt)
-        return receipt
-        
-
-        # logger.info(receipt)
-
-        # return 0
-
-
-    def checkAccess(self, user, researchpurpose):
-        hasAccess = web3.checkAccess(self, user, researchpurpose)
-        return hasAccess
-
-
-class LuceRegistry(models.Model):
-    contract_address = models.CharField(max_length=255, null=True)
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-
-    def deploy(self):
-
-        # account[0] as the administrator
-        result = luce_project.LUCERegistry.deploy({'from': accounts[0]})
-        if result.tx.status == 1:
-            self.contract_address = result.address
-            print("Deploy LUCERegistry contract succeeded")
-        else:
-            print("Deploy LUCERegistry contract failed")
-
-        return result.tx.status
-
-    def deploy_contract(self):
-        tx_receipt = web3.deploy_registry(self.user)
-        if type(tx_receipt) is list:
-            return tx_receipt
-        self.contract_address = tx_receipt["contractAddress"]
-        return tx_receipt
-
-    def is_registered(self, user, usertype):
-        if usertype == 'requester':
-            return self.is_registered_as_requester(user)
-
-    def is_registered_as_requester(self, user):
-        result = luce_project.LUCERegistry.at(self.contract_address).checkUser(
-            user.ethereum_public_key)
-        return result
-
-        # isregistered = web3.is_registered(self, user, usertype)
-        # return isregistered
-
-    def register_provider(self, user, estimate):
-        tx = web3.register_provider(self, user, estimate)
-        return tx
-
-    def register_requester(self, user, license, estimate):
-        logger.info(user.ethereum_private_key)
-        logger.info(user.ethereum_public_key)
-
-        sender = accounts.add(private_key=user.ethereum_private_key)
-        result = luce_project.LUCERegistry.at(self.contract_address).registerNewUser(
-            user.ethereum_public_key, 
-            license, 
-            {
-                'from': sender
-                }
-            )
-        print(result)
-        # tx = web3.register_requester(self, user, license, estimate)
-        return result
