+++ luce_vm/luce_django/luce/blockchain/admin.py
@@ -1,0 +1,3 @@
+from django.contrib import admin
+
+# Register your models here.
